BackEnd/ClickKYCWebhook/requirements.txt
pycryptodome
requests
fastapi
sqlalchemy
pymysql
uvicorn
debugpy
httpx
cryptography
python-dotenv
asyncmy
uvicorn[standard]
slowapi
pillow